+++ NEWS.md
@@ -1,3 +1,9 @@
+## QuantRegGLasso 0.3.0 (Release Date: 2024-01-08)
+### Overview 
+- Added a `predict` function for estimating top k coefficient functions via `qrglasso` object
+- Added a helper function to meet the pre-conditions of `predict`
+
+
 ## QuantRegGLasso 0.2.0 (Release Date: 2024-01-06)
 ### Overview 
 - Added one-dimensional example code in the main function `qrglasso`.
@@ -6,6 +12,6 @@
 ---
 ## QuantRegGLasso 0.1.0 (Release date: 2024-01-05)
 #### Overview 
-- Add a helper function to orthogonize B-splines
+- Added a helper function to orthogonize B-splines
 
 ---